% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_rds.R
\name{csv_rds}
\alias{csv_rds}
\title{Title}
\usage{
csv_rds(datafile, outputname)
}
\arguments{
\item{datafile}{a datafile that will be converted to a dataframe and then be written to a ".csv" file and a ".rds" file.}

\item{outputname}{the name of the output file. Needs to be characteristic ("between bracktes")}
}
\value{
a ".csv" file and a ".rds" file in a map called "csv_rds".
}
\description{
Title
}
\examples{
csv_rds(flu, "flu_dataset")
csv_rds(dengue, "dengue_data")
}
