% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainfo.R
\name{datainfo}
\alias{datainfo}
\title{Get info about a dataset}
\usage{
datainfo(dscolomn)
}
\arguments{
\item{dscolomn}{colomn of a dataset that needs to be inspected}
}
\value{
sum of the data, minimal value, maximal value, mean value, and sd value
}
\description{
Get info about a dataset
}
\examples{
dataset <- data.frame(test_object=(1:100),
                     weight=(rnorm(100, mean = 80, sd = 6)))
datainfo(dataset$weight)
}
