\name{dengue}
\alias{dengue}
\docType{data}
\title{
cases of dengue
}
\description{
cases of dengue around the world
}
\usage{dengue}
\format{
  A data frame with 6590 rows on the following 6 variables.
  \describe{
    \item{X}{row}
    \item{year}{year}
    \item{month}{month}
    \item{day}{day}
    \item{country}{country}
    \item{dengue_cases}{dengue cases}
  }
}

\source{
https://github.com/DataScienceILC/tlsc-dsfb26v-20_workflows
}

\keyword{datasets}
