\name{edcc_data}
\alias{edcc_data}
\docType{data}
\title{
COVID-19 related cases and deaths
}
\description{
COVID-19 related cases and deaths
}
\usage{edcc_data}
\format{
  A data frame with 28729 rows on the following 11 variables.
  \describe{
    \item{dateRep}{dataRep}
    \item{day}{day of the month}
    \item{month}{a numeric vector}
    \item{year}{a numeric vector}
    \item{cases}{a numeric vector}
    \item{deaths}{a numeric vector}
    \item{countriesAndTerritories}{country}
    \item{geoId}{geoId}
    \item{countryterritoryCode}{country code}
    \item{popData2020}{popData2020}
    \item{continentExp}{continentExp}
  }
}

\source{
https://www.ecdc.europa.eu/en/publications-data/data-daily-new-cases-covid-19-eueea-country
}

\keyword{datasets}
