\name{flu}
\alias{flu}
\docType{data}
\title{
Cases of the flu
}
\description{
cases of the flu
}
\usage{flu}
\format{
  A data frame with 19111 observations on the following 6 variables.
  \describe{
    \item{X}{row}
    \item{year}{year}
    \item{month}{month of the year}
    \item{day}{day of the month}
    \item{country}{country}
    \item{flu_cases}{cases of flu}
  }
}

\source{
https://github.com/DataScienceILC/tlsc-dsfb26v-20_workflows
}

\keyword{datasets}
