% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{Making a line_graph}
\usage{
line_graph(datafile, X, Y)
}
\arguments{
\item{datafile}{files that's gonna be used to make the line graph}

\item{X}{value on the x-axis}

\item{Y}{value on the y-axis}
}
\value{
a line graph
}
\description{
Making a line_graph
}
\examples{
line_graph(edcc_data, day, cases)
}
