% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_value.R
\name{q_value}
\alias{q_value}
\title{Calculate Q value}
\usage{
q_value(x)
}
\arguments{
\item{x}{numeric value}
}
\value{
Q value of the numeric value
}
\description{
Calculate Q value
}
\examples{
q_value(10)
q_value(86)
}
